fix(ros2): stop reading PhysX telemetry before it exists
GetVehicleTelemetryData dereferences the PhysX vehicle, which is null while the
hero is still initializing, so publishing VS-04 crashed the server. Drop the
wheel angular velocity fields and leave their validity bits clear, and delete a
map-publish guard whose macro is never defined.

Author: Jisang Yun <jisangyun@hanyang.ac.kr>

+++ LibCarla/source/carla/ros2/ROS2.cpp
@@ -219,9 +219,9 @@ void ROS2::PublishHmcFeedback(
       lng_op_mode,
       lat_op_mode,
       actuator_fault,
-      env_lng_ready == 1u ? lng_ctrl_ready : 0u,
-      env_lat_ready == 1u ? lat_ctrl_ready : 0u,
-      env_gear_ready == 1u ? gear_sel_ready : 0u,
+      static_cast<uint8_t>(env_lng_ready == 1u ? lng_ctrl_ready : 0u),
+      static_cast<uint8_t>(env_lat_ready == 1u ? lat_ctrl_ready : 0u),
+      static_cast<uint8_t>(env_gear_ready == 1u ? gear_sel_ready : 0u),
       stop_hold_ready};
 }
